Move ExportListenerAction code into the ExportPanel
